+++ CHANGELOG.md
@@ -10,6 +10,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Delete variants by type of analysis (wgs|wes|panel)
 - Links to cBioPortal, MutanTP53, IARC TP53, OncoKB, MyCancerGenome, CIViC
 ### Fixed
+- Deleted variants count
 ### Changed
 - Print output of variants delete command as a tab separated table